+++ src/pages/type/type.ts
@@ -1,3 +1,4 @@
+import { LoggerService } from './../../services/log4ts/logger.service';
 import { TypeModel } from './../../models/cellar.model';
 import { SearchPage } from './../search/search';
 import { TranslateService } from '@ngx-translate/core';
@@ -8,16 +9,18 @@ import { PouchdbService } from './../../services/pouchdb.service';
 import { AlertService } from './../../services/alert.service';
 
 @Component({
-  selector: 'page-typen',
+  selector: 'page-type',
   templateUrl: 'type.html'
 })
 export class TypePage {
 
   public type:TypeModel = new TypeModel("","");
   public typeList:Array<TypeModel>=[];
+  public typesMap:Map<any,any>
   public submitted:boolean;
   public typeForm:FormGroup;
   public newType:boolean=false;
+  public list:boolean=true;
   
   constructor(public navCtrl: NavController,
               public navParams: NavParams, 
@@ -25,36 +28,50 @@ export class TypePage {
               public formBuilder: FormBuilder,
               public alertService:AlertService,
               public alertController:AlertController,
-              public translate:TranslateService) {
+              public translate:TranslateService,
+              public logger:LoggerService) {
       this.typeForm = formBuilder.group({
-        type: ['',Validators.required]    
-    });
+        nom: ['',Validators.required]    
+    },{validator:this.noDouble.bind(this)});
     this.submitted = false;
-          
+    this.pouch.getDocsOfType('type').then(types => {
+        this.typesMap = new Map(types.map((el) => [el.nom, el]));
+        this.typeList = types;
+        this.logger.log("[TypePage - ionViewDidLoad]typeList : " + JSON.stringify(this.typeList));
+      });          
   }
 
   public ionViewDidLoad() {
-    console.log("[TypePage - ionViewDidLoad]called");
+    this.logger.log("[TypePage - ionViewDidLoad]called");
     // If we come on this page on the first time, the parameter action should be set to 'list', so that we get the see the list of appellations 
     if (this.navParams.get('action')=='list') {
-      this.pouch.getDocsOfType('type')
-      .then(result => { this.typeList = result;
-                        console.log("[TypePage - ionViewDidLoad]typeList : "+JSON.stringify(this.typeList));        
-                      });   
+      this.list = true;
     } 
     // if we come on this page with the action parameter set to 'edit', this means that we either want to add a new appellation (id parameter is not set)
     // or edit an existing one (id parameter is set)
     else {
+      this.list = false;
       if (this.navParams.get('id')) {
         this.pouch.getDoc(this.navParams.get('id'))
         .then(type => {
             this.type = type;
-            console.log("[AppellationPage - ionViewDidLoad]AppellationPage loaded : "+JSON.stringify(this.type));
+            this.logger.log("[AppellationPage - ionViewDidLoad]AppellationPage loaded : "+JSON.stringify(this.type));
         });
       } else 
         this.newType = true;
     }
   }
+
+  private noDouble(group: FormGroup) {
+    this.logger.debug("[Type.noDouble]nodouble called");
+    if(!group.controls.nom) return(null);
+    let testKey = group.value.nom;
+    if (this.typesMap && this.typesMap.has(testKey)) {
+      this.logger.log("[Type.noDouble]double detected");
+      return({double:true});
+    } else
+      return(null);
+  }    
 
   public editType(type) {
     if (type._id)
@@ -87,33 +104,48 @@ export class TypePage {
 }
 
  public deleteType() {
-  let alert = this.alertController.create({
-      title: this.translate.instant('general.confirm'),
-      message: this.translate.instant('general.sure'),
-      buttons: [
-        {
-          text: this.translate.instant('general.cancel'),
-          role: 'cancel',
-          handler: () => {
-            console.log('Cancel clicked');
-          }
-        },
-        {
-          text: this.translate.instant('general.ok'),
-          handler: () => {
-              this.pouch.deleteDoc(this.type).then(response => {
-                  if (response.ok) {
-                      //this.pouch.loadAppellationsRefList();                    
-                      this.alertService.success(this.translate.instant('type')+this.translate.instant('isDeleted'),SearchPage,'');
-                    } else {
-                      this.alertService.error(this.translate.instant('type')+this.translate.instant('isNotDeleted'),undefined,'');                        
+  let used=false;
+  if (this.type._id) {
+    this.pouch.getDocsOfType('vin')
+    .then(vins => {
+        vins.forEach(vin => {
+          if (vin.type._id == this.type._id)
+            used = true;
+        })
+        if (!used) {
+          let alert = this.alertController.create({
+              title: this.translate.instant('general.confirm'),
+              message: this.translate.instant('general.sure'),
+              buttons: [
+                {
+                  text: this.translate.instant('general.cancel'),
+                  role: 'cancel',
+                  handler: () => {
+                    this.logger.log('Cancel clicked');
                   }
-              });
-          }
+                },
+                {
+                  text: this.translate.instant('general.ok'),
+                  handler: () => {
+                      this.pouch.deleteDoc(this.type).then(response => {
+                          if (response.ok) {
+                              //this.pouch.loadAppellationsRefList();                    
+                              this.alertService.success(this.translate.instant('type')+this.translate.instant('isDeleted'),SearchPage,'');
+                            } else {
+                              this.alertService.error(this.translate.instant('type')+this.translate.instant('isNotDeleted'),undefined,'');                        
+                          }
+                      });
+                  }
+                }
+              ]
+          });
+          alert.present();
+        } else {
+          this.alertService.error(this.translate.instant('type.cantDeleteBecauseUsed'),undefined,'dialog');
         }
-      ]
-  });
-  alert.present();
+      }
+    );
+  }
 }
 
   private cleanValidatorModelObject(obj) {
